SERVER-95104 Implement an ephemeral queryable server (#27361)
This is an HTTP server that `mongod` interacts with when using
`--storageEngine=queryable_wt`. This storage engine is WiredTiger
running with custom file system handlers to perform calls over the
network.

Example usage
```
python3 buildscripts/resmokelib/testing/queryable_server.py --port 8000 --dbpath /data/db --verbose 1
```

The plan is to create a new suite (`queryable`) that runs this Python
server as a `hook` to create integration tests.