SERVER-93311: delete featureFlagCommonQueryFramework (#27430)
This PR is part of our effort to remove the Bonsai query optimizer,
which was guarded behind featureFlagCommonQueryFramework. With the
exception of the "analyze" command, the Bonsai entry points have been
removed from the code already, so we no longer need the feature flag to
guard them. For now, we'll make the "analyze" command test-only, since
we may re-use this code for the Cost-based Plan Ranker work.