SERVER-85808 Rename [in|from]Mongos members in ExpressionContext into [in|from]Router (#27288)
Aggregation code uses ExpressionContext inMongos/fromMongos members to
see if an aggregate is running in a router-role context or was sent by a
router. Now that mongod can act as a router in addition to mongos, the
members have been updated to inRouter/fromRouter to reflect that.

SERVER-94989 is the follow up to update the code around aggregation
command request too.