SERVER-94592 Check inner arguments for explain (#27373)
This PR addresses the security vulnerability in
[CVE-138](https://jira.mongodb.org/browse/CVE-138), where mongos can
crash when generic arguments are set incorrectly in the explain. The
most notable example is when the `uid` is improperly set in the `lsid`,
due to a failure to perform a privilege check for the user to set the
uid of an lsid. The authorization check is performed on the command
being explained, but not on the generic arguments when they're unwrapped
into an explain command. To set a different `uid` in the `lsid`, a user
must have the `impersonate` `ActionType`. Otherwise, the uid must be the
[same](https://github.com/10gen/mongo/blob/bd72cc59151b964d373886c4380a620dd90783b1/src/mongo/db/session/logical_session_id_helpers.cpp#L166-L181)
as the user's user digest.

The `wrapAsExplain` command will prune out generic arguments in the
inner command invocation instead of propagating them to the top-level
explain command. Any user who wants to set generic arguments will be
able to set them in the top-level command if they are authorized to set
the argument (primarily, to change the `uid` value of an `lsid`).