SERVER-95244 fix tassert() failure for upsert run against direct shard connection (#27508)
When an upsert runs against a direct shard connection and resolves to an
insert, the following sequence of operations is possible:
* Assume that the filtering metadata on the shard is stale.
* Since the operation is unversioned, it is able to establish an
execution plan without triggering a `StaleConfig` error.
* The upsert stage records the _id of the document it intends to insert,
and then tries to perform the insert by calling
`collection_internal::insertDocument()`.
* The insertion may fail (with the associated WUOW aborting) by throwing
`StaleConfig`. This happens because writes need to consult filtering
metadata to tag writes to orphan documents in the oplog.
* However, the upsert stage fails to undo the changes to the
`SpecificStats` which indicate that an upsert occurred. This leads to
the downstream tassert() failure, since the downstream code believes
that an insert has occurred when in fact it has not.

This simple fix simply changes the order of operations in the
`UpsertStage`. Before the fix, we updated the `SpecificStats` and then
called `collection_internal::insertDocument()`. After the fix, we do so
in the opposite order so that the stats are not changed if the insert
fails.