SERVER-94503 Read executable's binary name from procfs for stacktraces on Linux. (#27477)
From the man page for dl_iterate_phdr, we see that the record for the
"main program" (loaded by the OS loader) will have an empty name. This
patch reads the program name from "/proc/self/exe" if possible, and
leaves the name blank if that fails (matching prior behavior).

---------

Co-authored-by: Billy Donahue <BillyDonahue@users.noreply.github.com>