SERVER-95107 Pass client when creating AuthorizationSession (#27540)
This is a straightforward API change where the AuthorizationSession
takes a hold of a pointer to its owning client rather than the
AuthorizationManager.

Now that the we have multiple AuthorizationManagers because of Invisible
Sharding, the pointer held by the AuthorizationSession can go stale, so
it is better to grab the correct manager from the owning client.

---------

Co-authored-by: Shreyas Kalyan <shreyas.kalyan@mongodb.com>