SERVER-93579: Split jstests into component directories for jstests/concurrency/fsm_workloads (#26143)
[SERVER-93579](https://jira.mongodb.org/browse/SERVER-93579)
This PR moves files to subdirectories to help organize the
`concurrency/fsm_workloads` directory. The subdirectories were made by:
- Looking at similar files in other directories and seeing how they were
grouped in subdirectories
- Making directories based on some code-ownership

This PR is a mechanical change -- changes in files were for references.
CODEOWNERS were only added for StorEx OR if requested to be changed in
the PR review comments.
There are comments outlining some of the reasoning for why files were
grouped together/why the subdirectories were made.