SERVER-95302: Don't check return code in test that races shutdown (#27560)
# Issue

SERVER-95302

# Description

This test code has an expectation that races against mongod shutdown. We
lose the race if shutdown occurs before the command reply "i will not
compelete because we're shutting down" gets sent.

In https://jira.mongodb.org/browse/BF-35167 we began to lose that race
on slower evergreen targets. Whilst it is not clear what changed to
cause us to start losing that race a lot, we should still change the
test to avoid the race.

In this case, it doesn't matter *why* the command exits. We have other
failpoints/expectations for enforce that the index build progressed to a
certain point.