SERVER-92904 Consider PBRT size when creating cursor response batch (#27614)
Before this change the 'getMore' could fail with 'BSONObjectTooLarge',
because the size of the post-batch resume token was not considered when
creating the cursor response object. The probability of this error was
increasing with the size of the resume-token which depends on the size
of the '_id' field of the document.