SERVER-95385: Add more logging for rare test failure in BF-34783 (#27649)
* In BF-34783, a `checkpointInTheMiddle` test failed one time in
evergreen because the checkpoint write directory was not found in the
`flushUntilStopped` thread.
* This PR just adds some logging to make this problem easier to debug
when it reproes again. As far as I can tell, the corresponding test
should always have at least one checkpoint in that directory, and other
parts of the test already correctly assert that. So this log should make
debugging a bit easier if this reproes again.