SERVER-94165 Merge remote command response structs into one (#27601)
This commit merges the RemoteCommandResponse struct with
RemoteCommandResponseBase. Every RemoteCommandResponse now has a host
and port number (target) associated with it so this commit mainly
consists of updating each RemoteCommandResponse to make sure there is a
target. Since most of the RCRs are in tests and it would be a much
larger refactor to update these with a correct target, they will use a
test only private constructor to construct without a target.