SERVER-95142 rename field name meta to mFld in test (#27652)
This patch renames a field name from 'meta' to 'mFld' in
move_timeseries.js.
The test used field name 'meta' for the time-series collection's
metaField.
For a more thorough coverage that field name should be something else,
because internally the bucket format always uses 'meta' to store the
metadata,
so also using 'meta' as the user-provided name means that we would not
get
coverage of the translation.