SERVER-95075 Use `_id` predicates for update/delete when safe (#27449)
This fixes SERVER-95075 which resolves a performance BF.

My fix for SERVER-73641 made it so that we no longer generate `_id`
predicates for updates and deletes to time series collections. We had
correctness issues here since previously we would use `_id` predicates
even in the presence of extended range dates and the fix for
SERVER-73641 addressed this. However, it also caused a performance
degradation for ranged deletes.

This PR updates the update and delete path to generate `_id` predicates
when we know that there are no extended range dates. I also added a test
for updates and deletes and extended range dates. There is currently no
coverage for this from what I can tell.

As part of this I added a call to `MatchExpression::normalize()`, to the
generated bucket predicates, which had the effect of reordering them. I
had to update the existing tests accordingly to get them to pass.