SERVER-95240 Remove PeerID from SSL handshakes (#27634)
Previously, _ssl.get() was used as the peer ID in the SSLSetPeerID()
call. This caused the function to use the first bytes of the internal
SSL context struct as the peer ID, not the _ssl pointer address, leading
to the same peer ID being assigned to different SSL contexts. As a
result, the session cache reused sessions across different connections,
which should have been treated separately.