SERVER-94856 Support type-bracketed intervals (#27389)
This commit supports type-bracketed intervals which are built from
IndexBoundBuilder for queries. For example, query `{a: {$gte: "abc"}}`
will be translated to interval `["abc", {})` that has two bounds with
different types (string and object).

To support this, this PR relaxes the check in
`estimateIntervalCardinality()` (and `canEstimateInterval()`) where it
was strictly requiring both bounds to be of the same type. Instead, we
allow an interval if the interval is type-bracketed where the end bound
is the minimum value of the next type of the start bound.

The changes include:
* Adds `sameTypeBracketedInterval()` and `getMinMaxBoundForSBEType()` in
stats/value_utils.h.
* Removes dependency on the Bonsai headers `bound_utils.h`.
* Allows `estimateIntervalCardinality()` and `canEstimateInterval()` to
accept type-bracketed intervals.
* Removes `canEstimateBound()` in canEstimateInterval() for now until we
support type counts.
* Reuses a few methods in `IndexBoundsBuilderTest` so as to test
translated interval in value_utils_test.cpp.