SERVER-94691 Made "_configsvrSetClusterParameter" command to support retries when "previousTime" parameter is set (#27107)
The patch includes the following changes:

- making `_configsvrSetClusterParameter` command request to re-execute
if the "previousTime" parameter does not match "clusterParameterTime" on
the persisted cluster-wide parameter document but the value to set is
equal to the value currently stored (which would be the case if the
request execution is retried. The mongos may retry the command on the
config server if the command fails after changing the cluster-wide
parameter state but before returning a result ).
- refactoring of the update conflict detection logic into
`detectConcurrentUpdateConflict()` and
`_detectConcurrentUpdateConflict(`) methods.

---------

Co-authored-by: Mindaugas Malinauskas <mindaugas.malinauskas@mongodb.com>
Co-authored-by: Pol Piñol Castuera <pol.pinol@mongodb.com>