SERVER-87960 Update ssl_linear/*.js (#27175)
The original ticket suggested updating all files in ssl_linear to use
the `new Mongo()` API. However, since all of these tests (except for
`ssl_cert_selector.js`) rely on an environment variable set in a new
process to allow the SSL library to find the certificate store, it is
not feasible to use `new Mongo()` in these cases, as `new Mongo()` does
not spawn a new process.

I have only updated `ssl_cert_selector.js` because it is the only test
that does not rely on an environment variable set in a new process.