SERVER-95519 Increase timeout on tasks that invoke Bazel (#27803)
Bazel tasks will invoke bazel build //src/... which depending on remote
execution contention, may take a long time (especially in the case of
nightlies where many builds are kicked off at once)

Increase the timeout on these tasks for now to mitigate the issue.

The long term fix likely involves something along the lines of a pre-run
task attached to any bazel-dependent task with a long timeout that will
mostly guarantee a high cache hit ratio and reduce noise for teams with
tasks that depend on Bazel in the case of remote executor contention.
There are tradeoffs here though, so let's reassess once we have
everything in Bazel and have cleared up configuration issues that are
causing cache misses.