SERVER-91869: Fulfill promises outside of lock in ConnectionPool (#25639)
SERVER-41318 made it so ConnectionPool gives its callers back
SemiFutures and expects it won't have to run continuations inline when
it fulfills promises - so it fulfills them under its lock. But, since
callers sometimes chain ExecutorFutures running on a
NetworkInterfaceThreadPool onto those SemiFutures, it's still forced to
sometimes run continuations inline . This is because
NetworkInterfaceThreadPool runs things inline when its schedule method
is called from a network interface thread. This is itself a potential
problem - it's generally expected that fulfilling a promise associated
with only a SemiFuture won't require you to run continuations inline.

To address the immediate problem, this change makes it so ConnectionPool
fulfills promises outside of its lock. Due to the design of
ConnectionPool, this means a lot of the members of
ConnectionPool::SpecificPool take locks as function parameters now. For
the most part, this undoes a lot of SERVER-41318 and restores the status
quo from before then.

Note that the potential deadlock this lock order inversion hints at
seems to be impossible. `isDoingNetworking` seems to create a critical
section that covers the two codepaths that take the locks in the
opposite order, so we can only ever be in one at a time. Nonetheless,
this change improves lock hygiene and lets us remove a suppression.

---------

Co-authored-by: Saman Memaripour <amirsaman.memaripour@mongodb.com>