SERVER-95580 make ninja restat all bazel outputs (#27859)
In [SERVER-95361](https://jira.mongodb.org/browse/SERVER-95361) we set
up a phony shim dependency on bazel outputs to prevent ninja from
reporting a huge list of outputs on a failed build. However this causes
incremental rebuild weirdness where in the same build ninja will not
think the bazel outputs changed because were not restated, and ninja
would not rebuild the dependent actions, until the next build when it
would then stat the outputs and notice things change.

This change makes the shim dependency enable restat so that the bazel
outputs will be checked within the same build.