SERVER-75941 Remove special FCV handling for audit config (#27390)
Special handling was required with the audit config between 7.0 and 8.0,
because in 7.1, we switched from having the audit config being stored on
the config server in `config.settings` to having it be a cluster
parameter. This required writing a bunch of code to handle 1. the change
in format of the audit config, 2. handling what to do on
upgrade/downgrade between these versions, 3. different binary versions
of router/config server/shard server. Now that 8.0 is the lastLTS
version, we can remove all of this handling.

Changes: 
- Remove support for the audit config having a `generation`
  - Remove `_getAuditConfigGeneration` command
  - Remove the `AuditConfigFormat`
- Make the `clusterParameterTime` field in the AuditConfig object
non-optional (`LogicalTime::kUninitialized` allows us to give no value)
- Remove the AuditConfigClusterParameter feature flag
- Remove set_audit_config_direct.js, which only runs when this is
disabled
- Remove special FCV upgrade/downgrade handling around the audit config
- Remove support for audit config stored inside `config.settings`
- Remove audit synchronize job and audit op observer, which were part of
the old audit config mechanism