SERVER-71806 Remove ChunkRange::extractKeyPattern method (#27849)
ChunkRange::extractKeyPattern can leads to error because it can mislead
engineers that it provides the shard key pattern, when in reality it
only infers it from the provided key, and cannot figure out if the shard
key is meant to be hashed or not. I propose to remove this method, and
instead use the more explicit `Helpers::inferKeyPattern`, which sets no
expectation a about returning the _shard key_.
I've also reviewed the current usages of `Helpers::inferKeyPattern`:
- [dataSize
command](https://github.com/mongodb/mongo/blob/9d0a1940f830fdb88454209ba62958faf4f9cf1e/src/mongo/db/commands/dbcommands.cpp#L367):
If no `keyPattern` is explicitly passed to the command, then we use
`inferKeyPattern`, which assumes a non-hashed shard key pattern from the
provided `min` and `max`. I think this is fine, because we cannot really
tell from the min and max values if the shard key is meant to be hashed
or not. This could be improved by looking at the existing indexes, but I
feel like this is not high-priority and is not the objective of this
ticket.
- The usage I added on this PR. Infers a non-hashed shard key pattern
from the provided range. Checks that min and max have the exact same
fields and order.