SERVER-87065: Re-enable time-series validation checks in testing (#27850)
This re-enables time-series validation checks that had been disabled in
testing when we made the change to switch time-series validation
failures from warnings to errors. The two checks re-enabled here are:

- That the timestamp embedded in the OID matches that of the control.min
field
- That the control.min.timestamp field is >= the observed minimum
timestamp, rounded down for granularity

For some tests, this required updated the manually created buckets that
had been created in testing to pass validation - for example, some
manually created buckets had embedded timestamps that did not match
their control.min field/the observed minimum timestamps.

In the process of re-enabling checks, I encountered two scenarios that
might require changes in the validations checks:

1. For buckets that have extended range measurements, we can't enforce
that the control.min timestamp matches the timestamp embedded in the
OID, so we skip the check in that case
2. For buckets that have their granularity changed, we can't enforce
that the control.min.timestamp field is <= the minimum measurement
rounded down to granularity. See below example:

- Bucket is created with granularity seconds. control.min.timestamp =
2024-10-07:10:03:00, observed min rounded = 2024-10-07:10:03:00
- We change granularity to minutes - control.min.timestamp =
2024-10-07:10:03:00, observed min rounded = 2024-10-07:10:00:00
- We change granularity to hours - control.min.timestamp =
2024-10-07:10:03:00, observed min rounded = 2024-10-07:00:00:00

I changed the comparison here to be that the rounded control.min
timestamp is <= the rounded actual timestamp. I remember we discussed a
potentially stricter check a long while back @dhly-etc , involving
control.min >= observed min timestamp - rounding seconds; does that
sound right? It's something I can look into if the current check is too
loose