SERVER-94790 Add unit tests for ServiceContext and Client (#27699)
This PR adds a number of unit tests for the basic functionality of
ServiceContext and Client. The Client tests are fairly minimal and
mostly focused on behavior when binding or releasing a client to or from
a thread. The ServiceContext tests mostly cover creating and destroying
Clients and OperationContexts and what side effects those activities
have. There are also tests related to ConstructorAction execution and
ServiceContext functionality like killing and delisting operations.