SERVER-95606 Link router transaction server status section into both mongos and mongod (#27892)
The router-role "transactions" ServerStatusSection is defined in the TU
router_transactions_server_Status.cpp, which isn't previously wasn't
linked into mongod. This means when the router-role code is running on
mongod, it can't find this section, while it can on mongos. This commit
ensures the section is available to router-role code regardless of what
binary it runs on.