SERVER-95423 Update expectations on SASL SCRAM client cache (#27746)
In the sharding_scram_cache jstest, we make the assumption that cache
misses are about equal to total cache entries. This is incorrect because
when overwrites happen, i.e. whenever a previously used HostAndPort is
connected to with a new set of presecrets/secrets, we will have a cache
miss, but the total number of entries will remain the same. Due to this,
this test is inherently unreliable. I'm replacing this jstest with a
unittest that more directly tests how the cache works and that it is
used correctly by the SaslSCRAMClientConversationImpl. Also adds extra
D5 logging so that we can track when cache hits/misses/etc happen
better.