SERVER-94176: Add pipelineVersion and summaryStats to checkpoint (#27747)
* This is a pre-req for https://github.com/10gen/mongo/pull/27189
* This PR adds `summaryStats` and `pipelineVersion` fields in the
checkpoint manifest and increments the checkpoint data version to 4.
* StreamManager, CheckpointCoordinator, CheckpointStorage changes to use
`summaryStats` for returning the cumulative summary stats for a
processor. This is necessary for the "edit stream processors" feature.
After a pipeline modification summary stats should be retained, but
per-operator stats are reset.
* This PR adds a version 4 compat test to kafka.js and changes an
important unit test to use LocalDiskCheckpointStorage instead of
InMemoryCheckpointStorage.