SERVER-93114 ThreadPoolTaskExecutor controls when to read the next exhaust response (#27842)
This commit updates NetworkInterface::startExhaustCommand to return an
ExhaustResponseReader, which allows ThreadPoolTaskExecutor to only begin
reading the next exhaust response after it has finished processing the
previous one. This fixes a number of bugs that have been observed in the
exhaust path when exhaust responses were received faster than the
executor could process them. The internal state tracking of
ThreadPoolTaskExecutor was refactored to make this change in execution
responsibility more natural.