SERVER-94175: Support basic stream processor modify workflows (#27189)
This PR adds basic mongostream support for modifying a stream
processor's pipeline, as discussed in [Technical Design: Edit Stream
Processors](https://docs.google.com/document/d/18gnxFtGt5WRFLckppuHiQ1OaHJNSW5Wywsa0JpSjJcg/edit#heading=h.1sf01n5a72ed).

This PR:
* Amend stream_ops to accept a pipelineVersion argument. This is now
stored in checkpoints.
* Add basic StreamManager support for starting a modified stream
processor. To determine that a processor has been modified, the
`pipelineVersion` in the start request is compared to the restore
checkpoint.
* Adds StreamManager support for validating a modify request. The
existing `startStreamProcessor.validateOnly=true` is amended for this.
To validate a modify request we use the userPipeline in the restore
checkpoint, and a new Planner method to validate the modify is allowed.
* After a modify, a stream processor's summary stats are retained, but
operator level stats are reset. To achieve this, a new `summaryStats`
field is added in the checkpoint manifest.