SERVER-95666 use relative bazel toolchain path (#27956)
In SERVER-95525 we switched scons to using the bazel toolchain, however
in evergreen the toolchain is downloaded into the mci directory which
changes frequently. Because the mci directory became part of the scons
command line scons effectively treated it as a different command
breaking the scons cache.

This change switches to using a relative path from the convenience link
that bazel creates in the project root.