SERVER-95712 Enable bazel dynamic scheduling and docker sandbox (#27983)
This change turns on dynamic scheduling on platforms where docker is
install and podman is not. Dynamic scheduling allows jobs to performed
locally and remote at the same time and which ever finishes first is
taken as the result. This prevent local builds from getting stuck in
long queues in the remote pool.

This also somewhat requires good hermiticity as you do not want to link
code compiled on two different systems. So this also enables local
docker sandbox to guarantee local and remote results would match.