SERVER-89857: Remove duplicate errors/warnings from Validation output (#27871)
# Issue

SERVER-89857

# Description

Validation reports index-specific errors/warnings in two places, both in
the `indexdetails` and in the top-level `errors`/`warnings` fields. This
is redundant and unnecessary.

We change the behaviour to report index-specific errors only in the
`indexDetails` field. To ensure that the user is aware to look for them,
we also add a synthetic error/warning in the respective top-level fields
directing them to look at the index fields.