SERVER-95627 Forwardize index bounds in collapseQuerySolution when IXSCAN scans backward (#27909)
Reverses the index bounds if a IXSCAN node scans backward. This is
required because IndexBoundsBuilder::unionize() assumes intervals in
ascending direction. Note that a shard key pattern is always ascending,
so a backward direction in IXSCAN implies descending direction.
Therefore, we can simply rely on 'isn->direction' without needing to
check and forwardize each field individually.