SERVER-94940 Add jstests for mongod and mongos startup and shutdown (#27881)
This change adds two jstests: one for mongod startup and shutdown and
one for mongos. Each test starts up a server with various options and
immediately shuts it down. Other jstests cover this functionality
extensively of course, but these are intended as small, targeted tests
that can be included in the SP smoke test suite to get coverage of an
important area that would otherwise not be well-covered by that suite.

One particular area in which we had no test coverage whatsoever was
around the --fork option - the mongo shell doesn't even properly handle
servers started with --fork. This change tests each scenario it covers
both with and without --fork too add coverage of that option. To support
these tests, a failpoint `shutdownAtStartup` has been added to mongos. A
similar failpoint with the same name aready exists in mongod.

These had to be jstests as unit tests were a very poor fit for covering
startup and shutdown. Using jstests ensures we link all of the
initializers and decorations and such. In addition, our death test
infrastructure simply isn't up to the task of handling tests like these
- death tests don't work properly with UNITTEST_HAS_CUSTOM_MAINLINE and
they don't work properly with death tests that are expected to exit
cleanly, both of which are things we'd need for these tests.