SERVER-95189 Remove authz_manager_external_state classes (#27882)
Notable changes:

1. **Deleted the AuthzManagerExternalState classes:** these have now
been replaced with the AuthorizationRouter on one end and the
AuthorizationBackend
2. **Made the functions on the AuthorizationBackend protected:** there
was no easy way to do this, because a lot of the functions from the
AuthorizationBackend needed to be called from the
`CmdUMCTyped<T>::Invovation` class which is nested and therefore cannot
be declared a friend class. Instead, I created a passthrough nested
class inside of AuthorizationBackendInterface called `CmdUMCPassthrough`
that exposed a passthrough API. It is only declared in
AuthorizationBackendInterface, it is defined in
user_management_commands.cpp which means it is inaccessible from any
other file.
3. **Set an initialize() function on the AuthorizationManagerFactory:**
this seemed like a logical place to keep the initialize function since
it a/ only needs to be called at startup and b/ can use inheritance to
initialize both LDAP and OIDC implementations.

Most changes are related to that.

---------

Co-authored-by: Varun Ravichandran <varun.ravichandran@mongodb.com>
Co-authored-by: Shreyas Kalyan <shreyas.kalyan@mongodb.com>