SERVER-95488 stale_config_error_for_direct_shard_operation.js: Repeat the test setup before retrying the transaction (#28002)
`stale_config_error_for_direct_shard_operation.js` expects to run insert
operations on a node without routing information by forcing a stepup of
a secondary. The test runs inside a `withRetryOnTransientTxnError`
block. If the function is retried due to a transient error, the second
time around a new primary would step up, which can be the original
primary, which has updated routing information, breaking the
expectations of the test. This PR repeats the setup of the test before
each time the transaction is retried.