SERVER-95333 introduce uses_getmore_outside_of_transaction and correct usages of requires_getmore in tests. (#27823)
This PR:
* Introduces the tag `uses_getmore_outside_of_transaction`. This tag is
meant to differentiate tests that use `getMore`s outside a transaction
even when run in the transactions in the transactions override suite
from tests that use `getMore`s within transactions.
* Removes the `requires_getmore` tag from tests that only call `getMore`
in either the `setup()` or `teardown()` functions.
* Adds `requires_getmore` and `uses_getmore_outside_of_transaction` to
suites that need it, and removes tests listed explicitly in exclude
lists.