SERVER-95739 Fix TargetOnRemoteCommandError test (#28061)
`TargetOnRemoteCommandError` asynchronously sends a command and then
immediately cancels it, but if the command gets cancelled before
scheduling is finished, the code we are trying to test never gets
executed. If the command does get scheduled, we hang on `future.wait()`
since the response never gets fulfilled.

This commit adds a wrapper around `sendCommand` to block until the
command finishes scheduling and makes sure that all ready work is
drained upon executor shutdown. In the future, we will want all
`AsyncRPCTestFixture` classes to use `scheduleCommand` instead of
`sendCommand`