SERVER-95029: Add minimal implementation of the  $externalAPI operator (#27769)
## Description

This PR adds a stripped-down implementation of the $externalAPI
operator. It can make HTTPS requests to an external api with some
caveats. It only parses the URL from the web API connection but does not
support any parsing of headers/parameters/any other option specced in
the TD.

Notably, the operator is missing some features that make it
production-ready and some of them are as follows:

- [SERVER-95481] Translating parsed user-provided configs into operation
options
- [SERVER-95031] Evaluate headers and parameters expressions using stage
input documents
- [SERVER-95032] Error handling for when headers/parameters evaluate to
unexpected types