SERVER-95736 Fix use-after-free in ConnectionPool::_get (#28024)
This change fixes a use-after-free in ConnectionPool::_get introduced in
SERVER-91869. The issue is that `pool` is a `shared_ptr&`, so when we
copy it into the continuation after releasing the lock, the original
`shared_ptr` can be deleted out from under us. To fix it, we copy the
`shared_ptr` while holding the lock, and pass the local copy into the
continuation's closure instead.