SERVER-95671: Remove BSON output generation from collection_validation::validate() (#28048)
# Issue

SERVER-95671

# Description

Previously, the "output" of validation was split inconsistently between
the results struct and the BSON output object, various bits of data were
persisted in one or the other output container at different times for
different reasons.

In this change, we consolidate output generation as follows:
- All parts of validation below collection_validation::validate()
persist results to the ValidateResults *only*, all reference to the BSON
output builder have been removed.
- An explicit conversion is used to put a ValidateResults into a BSON
object.

This should help to clarify the relationship between BSON output,
ValidateResults output, and the validation pass.