SERVER-95310 Fix perf regression with makeBsonObj() VM builtin function (#27573)
SERVER-94597 touched some of the code related to makeBsonObj() and it
caused some 15%-20% perf regressions for the
"ProjectionDottedField.SinglePathXXXComponentsNestedArraysOfSizeXXX"
benchmarks. When I investigated the specific cause of the regressions, I
found that the regressions were happening because differences in the
compiler's decisions around inlining functions.

This PR fixes the regression by: (1) moving all of the helper functions
related to makeBsonObj() to a single .h file; (2) moving the
makeBsonObj() builtin function and the makeObj() builtin function each
to their own separate .cpp file; and (3) adding a
`MONGO_COMPILER_ALWAYS_INLINE` annotation (if it wasn't already present)
to most of the aforementioned helper functions.

I did some local perf testing with the
"ProjectionDottedField.SinglePathXXXComponentsNestedArraysOfSizeXXX"
benchmarks and confirmed that this PR fixes all of the regressions.