SERVER-95858 Disable thin archives due to Bazel cache issues (#28129)
We have an issue with bad Bazel cache results when using thin archives.
The shared object (.so) files get built of of the static library (.a)
files. Thin archive files are basically pointers to the contained object
(.o) files on disk. Because they just contain pointers, if the contained
.o files change the .a file doesn't change. If the .a file doesn't
change, Bazel will just go get the resulting .so file from cache.