SERVER-95133 improve FixedString API (#27986)
This `FixedString` version has some enhancements and bugfixes that bring
it closer to being a subset of the `std::fixed_string` proposal.

I'm not giving it a `string_view`-like API, however. It's unnecessary
since we can always cast it to a `StringData` and call
`string_view`-like API functions on that.

I'm removing the `operator const char*` conversion. We should be using
`StringData` for safety and performance.
I'm also going into the code that was using `FixedString` as a `const
char*` and changing it to use `FixedString` as a `StringData` instead.
Also changing some adjacent `std::string_view` uses to `StringData`, as
`std::string_view` is banned.