SERVER-94829 Fix dropCollection behavior when dropping timeseries collections (#25564)
In `$out` to timeseries, the stage creates the target collection in two
steps: first rename of the temporary collection to the buckets
namespace, then creation of the view. These two steps can interleave
with `createCollection` (or `createView`) to the timeseries namespace,
which breaks the assumption of `dropCollection`. If that happens, `$out`
will fail when trying to create the view, and during cleanup, it will
leave the broken buckets collection in place.

In these circumstances, when the user tries to drop the broken buckets
collection, `dropCollection` will see that there is a collection in the
namespace of the view, and proceed to drop the collection, ironically
leaving the buckets collection (which was requested to be dropped in the
first place) untouched.

This PR rearranges the logic of `dropCollection` to correctly handle all
cases. It also refines the logic of `$out` during cleanup.