SERVER-93614: Change mongod to mongot communication to be pinning mode by default. (#26836)
[SERVER-93614](https://jira.mongodb.org/browse/SERVER-93614)

The pinning mode communication method has already been built and tested
by @gewa24. This PR makes pinning mode the default, and no pinning can
be specified explicitly if desired.

Currently there is a resmoke suite named
`search_pinned_connections_auth.yml` that is now changed to
`search_no_pinned_connections_auth.yml` as pinning mode will now be
tested by default in the 'search.yml' suite. So we continue to keep a
test suite to test the non-default behavior.

Note: we should not merge this change until
[SERVER-93583](https://jira.mongodb.org/browse/SERVER-93583) is
addressed as this is now a known issue that can happen in pinning mode.