SERVER-94599: implement $minMaxScalar window function for removable implementations (#27507)
[SERVER-94599](https://jira.mongodb.org/browse/SERVER-94599)

Introduces a new window function $minMaxScalar to the $setWindowFields
aggregation stage.

`$minMaxScalar'`s output follows the equation: `((xi - min(x)) / (min(x)
- max(x))) * (sMax - sMin)) + sMin`
Where:
- `xi` is the value of the current document being processed / output
- `min(x)` is the minimum value of the document in the current window
- `max(x)` is the maximum value of the document in the current window
- `sMin` is the minimum of the output domain (set to 0 if not specified)
- `sMax` is the maximum of the output domain (set to 1 if not specified)

The motivation for introducing this new window function is to use it
during the de-sugaring phase of the soon to be introduce `$scoreFusion`
phase.

**Note**: This PR only introduces support for configured window bounds
that necessitate removing documents from the window (and thus left
"unbounded" window bounds are unsupported). The reason for this is to
bound the scope of this already lengthy PR. Non-removable window bounds
will be supported in the follow up ticket SERVER-95229. Filling in the
logic for the scaffolded
class`WindowFunctionExecMinMaxScalarNonRemovable` will support this, and
is already installed in the PR.

**Note**: $minMaxScalar breaks suit from other window functions in an
important way. Previously all other window function only depended on the
values of the documents in the current window, however $minMaxScalar
also relies on the value of the "current" document being processed,
because it gives a relative score / percentage that this document is
between the current min and max of the window. The implication of this
is we now need to plumb in the value of the current document in from
`DocumentSourceSetWindowFields::doGetNext()`. Notice the modifications
to the interfaces of `WindowFunctionExec::getNext()` and
`WindowFunctionState::getValue()` which now optionally take in the
Document/Value of the current document. This parameter is passed in as a
`boost::optional` because now some window functions need the value, and
some do not. Note lots of one-liner changes that now pass in
`boost::none` to the current DocumentValue, where it is un-needed.

---------

Co-authored-by: Charlie Swanson <charlie.swanson@mongodb.com>